Introduce further ambiguity tests when accessing columns by name.
These tests are imperfect, as it should be noted that SQLite's idea of a column name is not defined unless the AS clause is used, see the C API documentation.

Furthermore, it is not entirely clear what the JavaDoc wants JDBC driver to do. Clearly other drivers are happy with names of the form "tablename.colname", but the spec says:

    columnLabel - the label for the column specified with
    the SQL AS clause. If the SQL AS clause was not
    specified, then the label is the name of the column

So what we should do is unclear, users should simply use the AS clause, but there's no reason not to act more like other JDBC drivers.

http://article.gmane.org/gmane.comp.db.sqlite.jdbc/707

from neuro159 at gmail.com

darcs-hash:20080511113614-0c629-a19c69ad06ceedacfa82d6006cb80b6dab77777c.gz